+++ test/Provider/cleanup.test.tsx
@@ -1,24 +1,13 @@
 import { render } from '@testing-library/react'
 import { Provider } from '../../src'
 
-it('cleanup provided value', () => {
-  const cleanUpFunc = jest.fn((value) => value)
-  const { unmount } = render(
-    <Provider source={'value'} cleanUp={cleanUpFunc}>
-      <div></div>
-    </Provider>,
-  )
-  unmount()
-  expect(cleanUpFunc.mock.calls).toHaveLength(0)
-})
-
 class A {}
 
 it('cleanup provided Create', () => {
   const instance = new A()
   const cleanUpFunc = jest.fn((value) => value)
   const { unmount } = render(
-    <Provider source={() => instance} cleanUp={(value) => cleanUpFunc(value)}>
+    <Provider source={() => instance} cleanUp={(data) => cleanUpFunc(data)}>
       <div></div>
     </Provider>,
   )