+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.3.0
+
+- Fix doc typo
+- Add `ctor` param for `ProviderProps`
+- Separate `ProviderProps` to disallow using `cleanUp` when use it with `source: T`
+
 # 1.2.0
 
 - Cleanup function of `Provider` only called for `create` source.